SERVER-24104 Replace stdx::duration with mongo::Duration.
The mongo::Duration type does overflow checking on casts and arithmetic.

This patch also moves DESTRUCTOR_GUARD out of assert_util.h into
destructor_guard.h in order to break an include cycle with duration.h.

+++ src/mongo/db/commands/getmore_cmd.cpp
@@ -357,7 +357,7 @@ public:
                 ctx.reset();
 
                 // Block waiting for data.
-                Microseconds timeout(curOp->getRemainingMaxTimeMicros());
+                Microseconds timeout(static_cast<int64_t>(curOp->getRemainingMaxTimeMicros()));
                 notifier->wait(notifierVersion, timeout);
                 notifier.reset();
